pyproject.toml
[project]
name = "fastapi-protobuf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "betterproto-rust-codec>=0.1.1",
    "betterproto[compiler]==2.0.0b7",
    "fastapi>=0.115.11",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "requests>=2.32.3",
    "uvicorn[standard]>=0.34.0",
]

[tool.ruff]

line-length = 120
exclude = [
    "google_pb/**",
    "google_grpc/**",
    "betterproto_pb/**",
]
output-format = "concise"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM812","D203","D213"]



